requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.28 